chore(endpoint): merge endpoint_auto/endpoint_cache into protocol_endpoint
endpoint_auto.go, endpoint_cache.go and their tests were split out of the
same feature (OpenAI endpoint resolution + auto-detection) that already
lives in protocol_endpoint.go. Four files for one cohesive concern was
unnecessary fragmentation — folded them back in, organized under section
banners (static resolution / cache / runtime auto-detection dispatch).

No code changes beyond the move; verified via full endpoint_auto_test.go
suite now inlined in protocol_endpoint_test.go.
